move file hashing function+fix large stack alloc
