+++ core/src/object/cas.rs
@@ -46,21 +46,3 @@ pub async fn generate_cas_id(path: PathBuf, size: u64) -> Result<String, io::Err
 
 	Ok(hex.to_string())
 }
-
-pub async fn full_checksum(path: &str) -> Result<String, io::Error> {
-	const BLOCK_SIZE: usize = 1048576;
-	//read file as buffer and convert to digest
-	let mut reader = File::open(path).await?;
-	let mut context = Hasher::new();
-	let mut buffer = [0; 1048576];
-	loop {
-		let read_count = reader.read(&mut buffer).await?;
-		context.update(&buffer[..read_count]);
-		if read_count != BLOCK_SIZE {
-			break;
-		}
-	}
-	let hex = context.finalize().to_hex();
-
-	Ok(hex.to_string())
-}
