+++ core/src/object/mod.rs
@@ -1,6 +1,7 @@
 pub mod cas;
 pub mod identifier_job;
 pub mod preview;
+pub mod validation;
 
 // Objects are primarily created by the identifier from Paths
 // Some Objects are purely virtual, unless they have one or more associated Paths, which refer to a file found in a Location
